Add utilities file and refactor simid_protocol
